+++ test/python/sigopttest/handlers/validate/validate_dict.py
@@ -3,8 +3,7 @@
 # SPDX-License-Identifier: Apache License 2.0
 import pytest
 
-from zigopt.api.errors import InvalidTypeError, InvalidValueError, MissingJsonKeyError, SigoptValidationError
-from zigopt.api.validate_schema import validate
+from zigopt.api.errors import InvalidTypeError, MissingJsonKeyError, SigoptValidationError
 from zigopt.handlers.validate.validate_dict import (
   Base64InputValidator,
   Base64OutputValidator,
@@ -445,64 +444,3 @@ class TestValidateMutuallyExclusiveProperties(object):
     with pytest.raises(TypeError) as error:
       key_present("abc", "def")
     assert str(error.value) == "Expected json_obj to be a mapping, received 'str'"
-
-
-class TestValidateSchema(object):
-  def test_validate_simple_array(self):
-    schema = {
-      "type": "array",
-      "items": {
-        "type": "object",
-      },
-    }
-    with pytest.raises(InvalidTypeError):
-      validate([{}, 2, {}], schema)
-    with pytest.raises(InvalidTypeError):
-      validate(["a"], schema)
-    validate([], schema)
-
-  def test_simply_nested_array(self):
-    schema = {
-      "type": "object",
-      "properties": {
-        "key_of_array": {
-          "type": "array",
-          "items": {
-            "type": "string",
-          },
-        }
-      },
-    }
-    with pytest.raises(InvalidTypeError):
-      validate({"key_of_array": [1]}, schema)
-
-  def test_deeply_nested_array(self):
-    schema = {
-      "type": "object",
-      "properties": {
-        "key_of_object": {
-          "type": "object",
-          "properties": {
-            "key_of_array": {
-              "type": "array",
-              "items": {"type": "array", "items": {"type": "integer"}},
-            }
-          },
-        }
-      },
-    }
-    with pytest.raises(InvalidTypeError):
-      validate({"key_of_object": {"key_of_array": [["a"]]}}, schema)
-
-  def test_unspecific_index_error(self):
-    schema = {"type": "array", "maxItems": 1}
-    with pytest.raises(InvalidValueError):
-      validate([{}, 2, {}], schema)
-
-  def test_nested_unspecific_index_error(self):
-    schema = {
-      "type": "object",
-      "properties": {"key_of_array": {"type": "array", "maxItems": 1}},
-    }
-    with pytest.raises(InvalidValueError):
-      validate({"key_of_array": [{}, 2, {}]}, schema)
